add JSON save system
Comparison done between reading JSON with pixelsForce and calculate pixels force on the fly.
Calculating is better/same for few planets and objects on the level
JSON is better otherwise.
JSON reading is chosen for pixels force
Generation of JSON for level is necessary before using it.
Function to generate JSON added